+++ src/django/api/serializers/__init__.py
@@ -2,5 +2,10 @@
 from .user import UserSerializer
 from .utility import UtilitySerializer
 from .state import StateIDSerializer
-from .boundary import BoundaryListSerializer, BoundaryDetailSerializer
+from .boundary import (
+    BoundaryListSerializer,
+    BoundaryDetailSerializer,
+    NewBoundarySerializer,
+)
 from .shape import ShapeSerializer
+from .reference_image import ReferenceImageSerializer